+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+* Hide PHP Notices, instead save them in the `error.log` file
+  * You can change the error.log file path with `PhelConfig::setErrorLogFile(str)`
+
 ## 0.10.1 (2023-05-12)
 
 * Fixed the `phel\repl\doc` function.